Spk refactor (#43)
DAF files were completely restructured. The underlying data structures in DAFs are a vector of values describing metadata, and a f64 matrix of raw data. A data structure was created to support this, called DafArray.

Previously this data was loaded into custom structs, with re-scaling of coefficients so that they were slightly more favorable for the units which NEOSpy uses. However in doing this we remove the ability to exactly reproduce the original data file.

The DafArray struct was then distributed throughout the SPK and PCK data structures, which were updated to use it.

Along the way some uses of isize and usize were switched to i32, which is the correct type for the contents of DAF files.

+++ src/neospy/rust/simult_states.rs
@@ -93,7 +93,7 @@ impl PySimultaneousStates {
 
     /// The reference center NAIF ID for this state.
     #[getter]
-    pub fn center_id(&self) -> isize {
+    pub fn center_id(&self) -> i32 {
         self.0.center_id
     }
